+++ PokemonDL/feu.h
@@ -4,13 +4,13 @@
 class CFire : public CPokemon
 {
 private:
-	int m_burn;
+	float m_burn;
 
 public:
 	CFire();
 	bool burn();
 
-	int getBurn();
+	float getBurn();
 
 	void setBurn(float burn);
 };
